network_generation: split into header and source
