+++ meson.build
@@ -7,6 +7,7 @@ incdir = include_directories('include')
 sources_seldon = [
   'src/simulation.cpp',
   'src/network.cpp',
+  'src/network_generation.cpp',
   'src/model.cpp',
   'src/models/DeGroot.cpp',
   'src/util/tomlplusplus.cpp'
